Contra-Duplicidade
customerID e Cpf mudaram de INT para LONG, ja que o INT não aceita cpfs que comecem com o numero 4 ou maiores.
Mudanças no front de testes para que um erro 500 seja identificado como duplicidade
A coluna cpf não aceita repetições, ao identificar uma ele retorna o erro 500 no front.

+++ src/main/java/br/com/cadastro/repository/FichaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.cadastro.domain.Ficha;
 
-public interface FichaRepository extends JpaRepository<Ficha, Integer> {
+public interface FichaRepository extends JpaRepository<Ficha, Long> {
 
 }
